+++ resources/helpers/services/veil-helpers.php
@@ -2,7 +2,6 @@
 
 use Bayfront\Bones\App;
 use Bayfront\Container\NotFoundException;
-use Bayfront\RouteIt\DispatchException;
 use Bayfront\Veil\FileNotFoundException;
 use Bayfront\Veil\Veil;
 
@@ -63,32 +62,4 @@ function get_view(string $file, array $data = [], bool $minify = false): string
 function view(string $file, array $data = [], bool $minify = false): void
 {
     echo get_view($file, $data, $minify);
-}
-
-/**
- * Echo compiled template file, or dispatch to fallback if not existing.
- *
- * Returned value is filtered through the "veil.view" hook.
- *
- * @param string $file (Path to file from base path, excluding file extension)
- * @param array $data (Data to pass to view in dot notation)
- * @param bool $minify (Minify compiled HTML?)
- *
- * @throws NotFoundException
- * @throws DispatchException
- */
-
-function view_or_fallback(string $file, array $data = [], bool $minify = false)
-{
-
-    try {
-
-        view($file, $data, $minify);
-
-    } catch (NotFoundException | FileNotFoundException $e) {
-
-        get_router()->dispatchToFallback();
-
-    }
-
 }